+++ scripts/watch.js
@@ -25,7 +25,7 @@ chokidar.watch(srcPath, {
     mkdirp.sync(path.dirname(dist));
     let transformd = false;
     let code = '';
-    if (file.endsWith('.js') && !/\/flow\//.test(file)) {
+    if (file.endsWith('.js') && !/\/(flow|views)\//.test(file)) {
       transformd = true;
       try {
         code = babel.transformFileSync(file, {